+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.container.tsx
@@ -16,7 +16,8 @@ import {
     updateEnrollmentDate,
     updateIncidentDate,
     useCommonEnrollmentDomainData,
-    useRuleEffects,
+    useEnrollmentScopeRuleEffects,
+    selectEnrollmentWidgetEffects,
 } from '../../common/EnrollmentOverviewDomain';
 import {
     deleteEnrollment,
@@ -28,7 +29,6 @@ import { useCoreOrgUnit } from '../../../../metadataRetrieval/coreOrgUnit';
 import { DataStoreKeyByPage, EnrollmentPageLayout } from '../../common/EnrollmentOverviewDomain/EnrollmentPageLayout';
 import { useProgramMetadata, useProgramStages } from './hooks';
 import { useNavigate, buildUrlQueryString, useLocationQuery } from '../../../../utils/routing';
-import { useFilteredWidgetData } from './hooks/useFilteredWidgetData';
 import { useLinkedRecordClick } from '../../common/TEIRelationshipsWidget';
 import {
     useEnrollmentPageLayout,
@@ -52,7 +52,6 @@ export const EnrollmentPageDefault = () => {
     const queryClient = useQueryClient();
     const { status: widgetEnrollmentStatus } = useSelector(({ widgetEnrollment }: any) => widgetEnrollment);
     const { enrollmentId, programId, teiId, orgUnitId } = useLocationQuery();
-    const { orgUnit, error } = useCoreOrgUnit(orgUnitId);
     const { onLinkedRecordClick } = useLinkedRecordClick();
     const {
         pageLayout,
@@ -63,13 +62,15 @@ export const EnrollmentPageDefault = () => {
         dataStoreKey: DataStoreKeyByPage.ENROLLMENT_OVERVIEW,
     });
 
-    const program = useTrackerProgram(programId);
     const {
         error: enrollmentsError,
         enrollment,
         attributeValues,
         readOnly: trackedEntityInactive,
+        ownerOrgUnitId,
     } = useCommonEnrollmentDomainData(teiId, enrollmentId, programId);
+    const program = useTrackerProgram(programId);
+    const { orgUnit: ownerOrgUnit, error: ownerOrgUnitError } = useCoreOrgUnit(ownerOrgUnitId);
 
     const onStatusToggleSuccess = useCallback(() => {
         dispatch(setTrackedEntityInactiveStatus(!trackedEntityInactive));
@@ -86,20 +87,21 @@ export const EnrollmentPageDefault = () => {
         https://dhis2.atlassian.net/browse/DHIS2-17574
     */
 
-    if (programMetaDataError || enrollmentsError) {
+    if (programMetaDataError || enrollmentsError || ownerOrgUnitError) {
         log.error(errorCreator('Enrollment page could not be loaded')(
-            { programMetaDataError, enrollmentsError },
+            { programMetaDataError, enrollmentsError, ownerOrgUnitError },
         ));
     }
 
-    const ruleEffects = useRuleEffects({
-        orgUnit,
+    useEnrollmentScopeRuleEffects({
+        enrollmentId,
+        orgUnit: ownerOrgUnit,
         program,
         apiEnrollment: enrollment,
         apiAttributeValues: attributeValues,
     });
 
-    const outputEffects = useFilteredWidgetData(ruleEffects);
+    const outputEffects = useSelector(selectEnrollmentWidgetEffects);
     const hideWidgets = useHideWidgetByRuleLocations(program.programRules);
 
     const onDeleteTrackedEntitySuccess = useCallback(() => {
@@ -185,10 +187,6 @@ export const EnrollmentPageDefault = () => {
         );
     }
 
-    if (error) {
-        return error?.errorComponent;
-    }
-
     return (
         <EnrollmentAccessProvider program={program} trackedEntityInactive={trackedEntityInactive}>
             <EnrollmentPageLayout
@@ -197,6 +195,7 @@ export const EnrollmentPageDefault = () => {
                 availableWidgets={WidgetsForEnrollmentPageDefault}
                 teiId={teiId}
                 orgUnitId={orgUnitId}
+                ownerOrgUnitId={ownerOrgUnitId}
                 program={program}
                 stages={stages}
                 events={enrollment?.events}
@@ -221,7 +220,6 @@ export const EnrollmentPageDefault = () => {
                 onUpdateEnrollmentStatus={onUpdateEnrollmentStatus}
                 onUpdateEnrollmentStatusSuccess={onUpdateEnrollmentStatusSuccess}
                 onUpdateEnrollmentStatusError={onUpdateEnrollmentStatusError}
-                ruleEffects={ruleEffects}
                 widgetEnrollmentStatus={widgetEnrollmentStatus}
                 onAccessLostFromTransfer={onAccessLostFromTransfer}
                 feedbackEmptyText={i18n.t('No feedback for this enrollment yet')}