+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/index.ts
@@ -21,7 +21,12 @@ export {
     rollbackEnrollmentEvents,
 } from './enrollment.actions';
 export { useCommonEnrollmentDomainData } from './useCommonEnrollmentDomainData';
-export { useRuleEffects } from './useRuleEffects';
+export {
+    useEnrollmentScopeRuleEffects,
+    selectEnrollmentHiddenAttributeIds,
+    selectEnrollmentHiddenProgramStageIds,
+    selectEnrollmentWidgetEffects,
+} from './useRuleEffects';
 export {
     EnrollmentAccessProvider,
     useEnrollmentAccessContext,